requirements.txt
numpy
grpcio
protobuf
docker